# Project/api.yaml
openapi: 3.0.3
info:
  title: Distributed Music Editor  - Advanced Sound Systems
  description: |-
    A firma Advanced Sound Systems (ASS) encontra-se a desenvolver uma aplicação de karaoke para músicos. Esta aplicação distingue-se por não só remover a voz das músicas, mas também remover instrumentos individuais, permitindo a um músico substituir a performance do músico original pela sua. Este novo serviço será disponibilizado online através de um portal web em que o músico pode fazer upload de um ficheiro de música, analisar os instrumentos que compõem a música, selecionar vários instrumentos e finalmente receber um ficheiro novo em que a música contém apenas esses instrumentos.
    
    A tarefa de separar uma música em várias pistas por instrumento é intensiva do ponto de vista de processamento pelo que a ASS contratou os alunos de Computação Distribuída para desenvolver um serviço online capaz de atender às necessidades da empresa com uma qualidade de experiência elevada para o utilizador final (identificação rápida dos instrumentos e construção do novo ficheiro sem o(s) instrumento(s) seleccionados). 
    
    Para esta tarefa deverão desenvolver todo o código necessário para a criação de um serviço Web que possa servir em paralelo múltiplos clientes de forma rápida e eficiente recorrendo a computação paralela (aqui conseguido através de paralelização em múltiplos processos/workers independentes que podem estar no mesmo computador ou não).

  version: 1.0.0
...
paths:
  /music:
    post:
      tags:
        - music
      summary: Adicionar nova música
      description: Submeter ficheiros de áudio em formato MP3. De seguida o sistema deverá partir o ficheiro áudio em trechos de duração mais curta que podem ser analisados paralelamente. Em resposta à submissão do ficheiro de áudio, o serviço deverá retornar um ficheiro JSON contendo um identificador do trabalho (music_id), e uma lista de identificadores das pistas de instrumentos (cada identificador deverá corresponder a um instrumento) assim como um link para o ficheiro de cada instrumento (permitindo ao utilizador fazer o download e reprodução do ficheiro).
      operationId: submit
      requestBody:
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Musica'
        '405':
          description: Invalid input
    get:
      tags:
        - music
      summary: Listar todas as músicas submetidas
      description: Listar todas as músicas submetidas
      operationId: listAll
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Musica'
  /music/{music_id}:
    post:
      tags:
        - music
      summary: Processar uma musica
      description: Requisitar o processamento da música <music_id> com apenas os instrumentos listados através do seu identificador num ficheiro JSON submetido. 
      operationId: process
      parameters:
        - name: music_id
          in: path
          description: ID da musica a procesar
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: integer
              example: [1,2,3,5]
      responses:
        '200':
          description: successful operation
        '404':
          description: Music not found
    get:
      tags:
        - music
      summary: Estado de processamento de uma musica
      description: Requisitar o estado de processamento de uma música, quando o estado é 100% a resposta deve incluir o url do ficheiro final.
      operationId: progress
      parameters:
        - name: music_id
          in: path
          description: ID da musica a procesar
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                oneOf:
                  - type: integer
                    example: 43
                  - type: string
                    format: uri
                    example: http://localhost/file/31231234
        '404':
          description: Music not found
...
components:
  schemas:
    Musica:
      type: object
      properties:
        music_id:
          type: integer
          format: int64
          example: 123345678
        name:
          type: string
          example: "Another One Bites the Dust"
        band:
          type: string
          example: "Queen"
        tracks:
          type: array
          items:
            $ref: '#/components/schemas/Track'
      xml:
        name: musica
    Track:
      type: object
      properties:
        name:
          type: string
          example: drums
        track_id:
          type: integer
          format: int64
          example: 1
...